Centralized Cache Timers, Fixed UI Bugs, Made JF Prices public
Added UI hint for corp and alliance prices on the jf admin page.
Applied fix for making sure code doesn't run if the "multiple init" box on the jf admin page is empty.
All cache timers have been removed from the cache collections and moved to teh "caches" collection.
Fixed nav bar errors when signed in, but not in alliance.
Made the JF page public, but must be signed in to see contract history.